Moved updateNftsPinningStatus to nfts endpoints
